refactors chat message formatting into component
